# packages/libs/sdk/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Agents Control Plane API",
        "version": "1.0.0",
        "description": "REST API for managing agents, jobs, triggers, and webhooks."
    },
...
    "paths": {
        "/agents": {
            "post": {
                "tags": ["Agents"],
                "description": "Create a new agent",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "minLength": 1
                                    },
                                    "activities": {
                                        "type": "array",
                                        "items": {
                                            "type": "object",
                                            "properties": {
                                                "type": {
                                                    "type": "string",
                                                    "minLength": 1
                                                },
                                                "id": {
                                                    "type": "string"
                                                },
                                                "params": {},
                                                "maxRetries": {
                                                    "type": "integer",
                                                    "minimum": 0
                                                },
                                                "onError": {
                                                    "type": "object",
                                                    "additionalProperties": {
                                                        "type": "string"
                                                    }
                                                }
                                            },
                                            "required": ["type"],
                                            "additionalProperties": false
                                        },
                                        "default": []
                                    },
                                    "failure_threshold": {
                                        "type": "integer",
                                        "exclusiveMinimum": 0,
                                        "default": 3
                                    }
                                },
                                "required": ["name"],
                                "additionalProperties": false
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Default Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "string"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "activities": {},
                                        "failure_threshold": {
                                            "type": "number"
                                        },
                                        "failure_count": {
                                            "type": "number"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "required": [
                                        "id",
                                        "name",
                                        "status",
                                        "failure_threshold",
                                        "failure_count",
                                        "created_at",
                                        "updated_at"
                                    ],
                                    "additionalProperties": false
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Default Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "type": {
                                            "type": "string"
                                        },
                                        "title": {
                                            "type": "string"
                                        },
                                        "status": {
                                            "type": "number"
                                        },
                                        "detail": {
                                            "type": "string"
                                        },
                                        "instance": {
                                            "type": "string"
                                        }
                                    },
                                    "required": ["type", "title", "status", "detail"],
                                    "additionalProperties": false
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": ["Agents"],
                "description": "List all agents",
                "responses": {
                    "200": {
                        "description": "Default Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "id": {
                                                "type": "string"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "status": {
                                                "type": "string"
                                            },
                                            "activities": {},
                                            "failure_threshold": {
                                                "type": "number"
                                            },
                                            "failure_count": {
                                                "type": "number"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "required": [
                                            "id",
                                            "name",
                                            "status",
                                            "failure_threshold",
                                            "failure_count",
                                            "created_at",
                                            "updated_at"
                                        ],
                                        "additionalProperties": false
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/agents/{id}": {
            "get": {
                "tags": ["Agents"],
                "description": "Get an agent by ID",
                "parameters": [
                    {
                        "schema": {
                            "type": "string"
                        },
                        "in": "path",
                        "name": "id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Default Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "string"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "activities": {},
                                        "failure_threshold": {
                                            "type": "number"
                                        },
                                        "failure_count": {
                                            "type": "number"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "required": [
                                        "id",
                                        "name",
                                        "status",
                                        "failure_threshold",
                                        "failure_count",
                                        "created_at",
                                        "updated_at"
                                    ],
                                    "additionalProperties": false
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Default Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "type": {
                                            "type": "string"
                                        },
                                        "title": {
                                            "type": "string"
                                        },
                                        "status": {
                                            "type": "number"
                                        },
                                        "detail": {
                                            "type": "string"
                                        },
                                        "instance": {
                                            "type": "string"
                                        }
                                    },
                                    "required": ["type", "title", "status", "detail"],
                                    "additionalProperties": false
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": ["Agents"],
                "description": "Delete an agent",
                "parameters": [
                    {
                        "schema": {
                            "type": "string"
                        },
                        "in": "path",
                        "name": "id",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Agent deleted"
                    }
                }
            }
        },
        "/agents/{id}/invoke": {
            "post": {
                "tags": ["Agents"],
                "description": "Invoke an agent to create a new job",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "payload": {
                                        "default": null
                                    }
                                },
                                "additionalProperties": true
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "schema": {
                            "type": "string"
                        },
                        "in": "path",
                        "name": "id",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Default Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "string"
                                        },
                                        "agent_id": {
                                            "type": "string"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "payload": {},
                                        "result": {},
                                        "error": {
                                            "type": ["null", "string"]
                                        },
                                        "current_step_id": {
                                            "type": ["null", "string"]
                                        },
                                        "context": {
                                            "type": "object",
                                            "additionalProperties": {}
                                        },
                                        "step_retries": {
                                            "type": "number"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "required": [
                                        "id",
                                        "agent_id",
                                        "status",
                                        "error",
                                        "current_step_id",
                                        "context",
                                        "step_retries",
                                        "created_at",
                                        "updated_at"
                                    ],
                                    "additionalProperties": false
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Default Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "type": {
                                            "type": "string"
                                        },
                                        "title": {
                                            "type": "string"
                                        },
                                        "status": {
                                            "type": "number"
                                        },
                                        "detail": {
                                            "type": "string"
                                        },
                                        "instance": {
                                            "type": "string"
                                        }
                                    },
                                    "required": ["type", "title", "status", "detail"],
                                    "additionalProperties": false
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Default Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "type": {
                                            "type": "string"
                                        },
                                        "title": {
                                            "type": "string"
                                        },
                                        "status": {
                                            "type": "number"
                                        },
                                        "detail": {
                                            "type": "string"
                                        },
                                        "instance": {
                                            "type": "string"
                                        }
                                    },
                                    "required": ["type", "title", "status", "detail"],
                                    "additionalProperties": false
                                }
                            }
                        }
                    }
                }
            }
        },
        "/agents/{id}/restart": {
            "post": {
                "tags": ["Agents"],
                "description": "Restart a paused agent",
                "parameters": [
                    {
                        "schema": {
                            "type": "string"
                        },
                        "in": "path",
                        "name": "id",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Agent restarted"
                    },
                    "404": {
                        "description": "Default Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "type": {
                                            "type": "string"
                                        },
                                        "title": {
                                            "type": "string"
                                        },
                                        "status": {
                                            "type": "number"
                                        },
                                        "detail": {
                                            "type": "string"
                                        },
                                        "instance": {
                                            "type": "string"
                                        }
                                    },
                                    "required": ["type", "title", "status", "detail"],
                                    "additionalProperties": false
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}